cleaned up the files and folders a bit
