+++ server/package.json
@@ -13,9 +13,11 @@
   "license": "ISC",
   "devDependencies": {
     "nodemon": "^3.1.7",
+    "swagger-autogen": "^2.23.7",
     "ts-node": "^10.9.2"
   },
   "dependencies": {
+    "@types/swagger-ui-express": "^4.1.6",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "mongoose": "^8.7.1",
